lib/foc : Use non-zero defaults for speed and torque
System fault is observed when speed_sp is set to zero.
Avoid this by having other small number as reset/default value.

+++ lib/src/foc.h
@@ -34,8 +34,8 @@ public:
 private:
 	IIO_Driver *mFoc_IIO_Handle;
 	static const std::string kFocDriverName;
-	double mTargetSpeed;
-	double mTargetTorque;
+	int mTargetSpeed;
+	int mTargetTorque;
 };
 
 #endif // _FOC_HPP_
